Update index css script
This commit makes changes to the index css and script files in order to display <div id=win> and <div id=lose> in the way I want it. When the player wins, it will hide the container div to only display the win div. When the player loses, it will hide the keyboard div to display the lose div while still displaying the container div with the hung stickman.

Other minor changes to style of the #visible-word div